api/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib
pydantic-settings
pydantic[email]
python-multipart
bcrypt
apscheduler
pandas
requests
duckdb
psycopg2
python-dotenv